Make it compile for ESP32-C3 but needs to be checked if things did not get broken

+++ src/log.cpp
@@ -369,9 +369,9 @@ void vTaskLOG(void* pvParameters)
 #ifdef WITH_SPIFFS
     { size_t Total, Used;
       if(SPIFFS_Info(Total, Used)==0)                            // get the SPIFFS usage summary
-      { Format_UnsDec(CONS_UART_Write, Used/1024);
+      { Format_UnsDec(CONS_UART_Write, (uint32_t)(Used/1024));
         Format_String(CONS_UART_Write, "kB used, ");
-        Format_UnsDec(CONS_UART_Write, Total/1024);
+        Format_UnsDec(CONS_UART_Write, (uint32_t)(Total/1024));
         Format_String(CONS_UART_Write, "kB total, "); }
     }
 #endif
